requirements.txt
# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0

# Data Storage
pymongo>=4.5.0

# LLM & Agent Framework
langchain>=0.1.0
langchain-core>=0.1.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Additional dependencies
motor>=3.3.0  # AsyncIO MongoDB driver
langchain-deepseek>=0.1.0  # Optional, for LLM-based classification
typing-extensions>=4.5.0
asyncio>=3.4.3
